Change DB to Postgres

+++ app/database/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean, Text, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, Boolean, Text, DateTime, ForeignKey
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 from datetime import datetime
@@ -9,7 +9,7 @@ Base = declarative_base()
 class User(Base):
     __tablename__ = 'users'
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)  # Изменено на BigInteger для Telegram ID
     lang = Column(String, default='ru')
     notify_time = Column(String, default='09:00')
     is_active = Column(Boolean, default=True)  # Для soft delete
@@ -23,8 +23,8 @@ class User(Base):
 class Debt(Base):
     __tablename__ = 'debts'
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey('users.user_id'), nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True)  # Обычный автоинкремент можно оставить Integer
+    user_id = Column(BigInteger, ForeignKey('users.user_id'), nullable=False)  # Изменено на BigInteger
     person = Column(String, nullable=False)
     amount = Column(Integer, nullable=False)
     currency = Column(String, default='UZS')
@@ -43,17 +43,18 @@ class Debt(Base):
 class ScheduledMessage(Base):
     __tablename__ = 'scheduled_messages'
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey('users.user_id'), nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True)  # Обычный автоинкремент можно оставить Integer
+    user_id = Column(BigInteger, ForeignKey('users.user_id'), nullable=False)  # Изменено на BigInteger
     text = Column(Text, nullable=False)
     photo_id = Column(String, nullable=True)
     schedule_time = Column(String, nullable=False)
     sent = Column(Boolean, default=False)
-    is_active = Column(Boolean, default=True) 
+    is_active = Column(Boolean, default=True)
     created_at = Column(DateTime, default=datetime.utcnow)
 
     # Relationships
     user = relationship("User", back_populates="scheduled_messages")
+
 
 def safe_str(value):
     """Безопасное преобразование в строку"""
